Change ctest removed, cmake fetch removed, google test update
The googletest cmake configuration is sadly slow, and the cmake are
configuring gtest under not required conditions. There is probably a way
to correctly configure this behavior in cmake. It's very complicated and
I don't feel like wasting my day trying to fix this stupid
configuration.
So, the googletest source are in the project dir. If path breaks
just look for just configure the gtest lib/include paths.
Googletest will be added to the git submodules soon.